USF-4151: Fix view/purchase history across multiple ACO storefronts
